requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
dataclasses-json==0.6.5
distro==1.9.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.17
langchain-community==0.0.37
langchain-core==0.1.52
langchain-openai==0.1.6
langchain-pinecone==0.1.0
langchain-text-splitters==0.0.1
langsmith==0.1.55
marshmallow==3.21.2
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.26.0
orjson==3.10.3
packaging==23.2
pinecone-client==3.2.2
pip==24.0
pydantic==2.7.1
pydantic_core==2.18.2
python-dotenv==1.0.1
PyYAML==6.0.1
regex==2024.4.28
requests==2.31.0
setuptools==69.5.1
sniffio==1.3.1
SQLAlchemy==2.0.30
tenacity==8.3.0
tiktoken==0.6.0
tqdm==4.66.4
typing_extensions==4.11.0
typing-inspect==0.9.0
urllib3==2.2.1
wheel==0.43.0
yarl==1.9.4
